requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
Flask-Babel==4.0.0
Flask-Session==0.8.0
reportlab==4.0.7
python-dateutil==2.8.2
Flask-Migrate==4.1.0
